+++ lib/api_connection/api_connection.dart
@@ -1,6 +1,6 @@
 class Api {
-  static const ip = '192.168.1.11';
-  // static const ip = 'localhost';
+  // static const ip = '192.168.1.11';
+  static const ip = 'localhost';
   static const hostConnect = 'http://$ip/attendifyyy_backend';
   static const signUp = '$hostConnect/authentication/sign_up.php';
   static const logIn = '$hostConnect/authentication/log_in.php';
@@ -17,4 +17,12 @@ class Api {
   static const retrieveImage = '$hostConnect/retrieve_image.php';
   static const deleteSchedule = '$hostConnect/create_schedule/delete_schedule.php';
   static const updateAccount = '$hostConnect/account/update_account.php';
+  static const getStudentData = '$hostConnect/create_student/student_data.php';
+  static const updateStudentData = '$hostConnect/create_student/update_student.php';
+  static const deleteStudent = '$hostConnect/create_student/delete_student.php';
+  static const deleteSubject = '$hostConnect/create_subject/delete_subject.php';
+  static const getSubjectData = '$hostConnect/create_subject/subject_data.php';
+  static const updateSubjectData = '$hostConnect/create_subject/update_subject.php';
+  static const updateSchedule = '$hostConnect/create_schedule/update_schedule.php';
+  static const getRecentAttendance = '$hostConnect/utils/recent_attendance.php';
 }
