+++ lib/core/dashboard.dart
@@ -2,23 +2,16 @@ import 'dart:convert';
 import 'package:attendifyyy/api_connection/api_connection.dart';
 import 'package:attendifyyy/attendance/attendance_list.dart';
 import 'package:attendifyyy/authentication/user_preferences/user_preferences.dart';
+import 'package:attendifyyy/create_schedules/create_schedule.dart';
+import 'package:attendifyyy/create_students/create_students.dart';
 import 'package:attendifyyy/create_subjects/create_subjects.dart';
-import 'package:attendifyyy/core/widgets/create_student_card.dart';
-import 'package:attendifyyy/core/widgets/stats_card.dart';
-import 'package:attendifyyy/core/widgets/list_page_nav_btn.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart';
-import '../schedules/create_schedule.dart';
 
-/*
-* current date is use for the current date below the dropdown sections
-* */
 String currentDate = DateFormat('EEEE, d MMM yyyy').format(DateTime.now());
 
 class DashboardScreen extends StatefulWidget {
-  const DashboardScreen({super.key});
-
   @override
   State<DashboardScreen> createState() => _DashboardScreenState();
 }
@@ -137,7 +130,7 @@ class _DashboardScreenState extends State<DashboardScreen> {
         child: AppBar(
           centerTitle: true,
           automaticallyImplyLeading: false,
-          backgroundColor: const Color.fromARGB(255, 255, 255, 255),
+          backgroundColor: Color.fromARGB(255, 255, 255, 255),
           title: Container(
             padding: const EdgeInsets.fromLTRB(4.0, 10.0, 4.0, 4.0),
             child: Image.asset(
@@ -151,11 +144,6 @@ class _DashboardScreenState extends State<DashboardScreen> {
       body: ListView(
         padding: const EdgeInsets.symmetric(vertical: 20.0, horizontal: 14.0),
         children: [
-          /*
-          *
-          * This is the headings that contains the teachers name, profile image, and current date
-          *
-          * */
           Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: [
@@ -193,11 +181,7 @@ class _DashboardScreenState extends State<DashboardScreen> {
             ],
           ),
           const SizedBox(height: 15),
-          /*
-          *
-          * Overview card
-          *
-          * */
+          //Overview
           Container(
             decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(14.0),
@@ -206,11 +190,7 @@ class _DashboardScreenState extends State<DashboardScreen> {
             padding: const EdgeInsets.all(16.0),
             child: Column(
               children: [
-                /*
-                *
-                * Overview headings row
-                *
-                * */
+                //Overview headings row
                 const Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
@@ -226,11 +206,7 @@ class _DashboardScreenState extends State<DashboardScreen> {
                 const SizedBox(
                   height: 14.0,
                 ),
-                /*
-                *
-                * Stats Row
-                *
-                * */
+                //Stats Row
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceAround,
                   children: [
@@ -246,31 +222,26 @@ class _DashboardScreenState extends State<DashboardScreen> {
             ),
           ),
           const SizedBox(height: 14.0),
-          /*
-          *
-          * attendance,class and students buttons section
-          *
-          * */
+          //attendance,class and students buttons section
           Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-            createListPageNavButton(context, const ListOfSubjects(), 0xFFCD00B9,
+            createListPageNavButton(context, ListOfSubjects(), 0xFFCD00B9,
                 Icons.subject, "Subject List"),
+            //spaces between buttons
             const SizedBox(height: 7),
             createListPageNavButton(
                 context,
                 const ListOfSchedules(),
                 0xFFFF9900,
                 Icons.schedule,
-                "Schedule List"),
+                "Schedule List"), //! kani na list of schedules styli na, okay na backend
+            //spaces between buttons
             const SizedBox(height: 7),
-            createListPageNavButton(context, const AttendanceReport(), 0xFF039000,
+            createListPageNavButton(context, AttendanceReport(), 0xFF039000,
                 Icons.group, "Attendance Report"),
+            //spaces between buttons
             const SizedBox(height: 30),
           ]),
-          /*
-          *
-          * recent attendance section
-          *
-          * */
+          //recent attendance section
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
@@ -281,11 +252,7 @@ class _DashboardScreenState extends State<DashboardScreen> {
                     fontSize: 18.0,
                   )),
               const SizedBox(height: 14.0),
-              /*
-              *
-              * student card list
-              *
-              * */
+              //student card list
               recentAttendanceList.isNotEmpty ? createStudentCard(recentAttendanceList[0]['attendance_status'] ?? "", '${recentAttendanceList[0]['first_name']}\n${recentAttendanceList[0]['last_name']}' ?? "", recentAttendanceList[0]['attendance_time'] ?? "") : createStudentCard("No data", "No data", "No data"),
             ],
           ),
@@ -293,4 +260,179 @@ class _DashboardScreenState extends State<DashboardScreen> {
       ),
     );
   }
+}
+
+/*
+ *
+ * Widget Components
+ *
+ * */
+
+//create statistic card such as present, absent, and late count.
+Widget createStatsCard(
+    IconData materialIcon, int backgroundColor, String label, int count) {
+  return Container(
+    width: 90.0,
+    decoration: BoxDecoration(
+      borderRadius: BorderRadius.circular(14.0),
+      //background color of stats card
+      color: Color(backgroundColor),
+    ),
+    padding: const EdgeInsets.all(7.0),
+    child: Column(
+      children: [
+        //stats card heading
+        Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(materialIcon, color: Colors.white),
+            Text(label,
+                style: const TextStyle(
+                  fontSize: 10.0,
+                  color: Colors.white,
+                  fontWeight: FontWeight.bold,
+                ))
+          ],
+        ),
+        //stats count
+        Text('$count',
+            style: const TextStyle(
+              color: Colors.white,
+              fontWeight: FontWeight.bold,
+              fontSize: 50.0,
+            ))
+      ],
+    ),
+  );
+}
+
+//create a listNavButton that navigate you to a certain list page
+Widget createListPageNavButton(
+    context, onClickAction, backgroundColor, materialIcon, label) {
+  return ElevatedButton(
+    onPressed: () {
+      Navigator.pushReplacement(
+        context,
+        MaterialPageRoute(builder: (context) => onClickAction),
+      );
+    },
+    style: ButtonStyle(
+        backgroundColor:
+            MaterialStateProperty.all<Color>(Color(backgroundColor)),
+        padding:
+            MaterialStateProperty.all<EdgeInsets>(const EdgeInsets.all(16.0)),
+        shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+            RoundedRectangleBorder(borderRadius: BorderRadius.circular(10.0)))),
+    child: Row(
+      children: [
+        Icon(
+          materialIcon,
+          color: Colors.white,
+          size: 35.0,
+        ),
+        const SizedBox(
+          width: 10,
+        ),
+        Text(label,
+            style: const TextStyle(
+                color: Colors.white,
+                fontWeight: FontWeight.bold,
+                fontSize: 16.0))
+      ],
+    ),
+  );
+}
+
+//create students card for recent attendance section
+Widget createStudentCard(status, name, timeIn) {
+  //default
+  int _iconColor = 0xFF039000;
+  var _materialIcon = Icons.playlist_add_check_outlined;
+
+  //set status icon
+  switch (status) {
+    case "Present":
+      {
+        _iconColor = 0xFF039000;
+        _materialIcon = Icons.playlist_add_check_outlined;
+      }
+      break;
+    case "Late":
+      {
+        _iconColor = 0xFFFF9900;
+        _materialIcon = Icons.timer_off_rounded;
+      }
+      break;
+    case "Absent":
+    default:
+      {
+        _iconColor = 0xFFFF9900;
+        _materialIcon = Icons.person_off;
+      }
+  }
+
+  return Container(
+    //This box decoration purpose is to make the edge of the student card a
+    //a bit round.
+    decoration: const BoxDecoration(
+      color: Colors.transparent,
+    ),
+    child: Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: <Widget>[
+        //container for image and name of the student that is place at the
+        // left side of the student card
+        Row(
+          children: [
+            //This is the first child of the left container.
+            //This is use to contain the student image
+            //but right now i use account_circle icon as substitute
+            Container(
+              margin: const EdgeInsets.only(
+                  right: 7), //margin between avatar and name
+              child: const Icon(
+                Icons.account_circle,
+                size: 50,
+                color: Colors.blueAccent,
+              ),
+            ),
+            //This is the second child of the left container.
+            //This is used to contain the student name
+            Text(
+              name,
+              style: const TextStyle(
+                  color: Colors.black,
+                  fontSize: 14,
+                  fontWeight: FontWeight.bold),
+            ),
+          ],
+        ),
+        //This is the container for time-in of student.
+        //This is located at the right side of the student card
+        Column(
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Icon(_materialIcon, color: Color(_iconColor)),
+                Text(status,
+                    style: TextStyle(
+                      color: Color(_iconColor),
+                      fontWeight: FontWeight.bold,
+                    ))
+              ],
+            ),
+            //This is the time in of the student
+            Text(
+              '$timeIn TIME IN',
+              style: TextStyle(
+                  color: Colors.black,
+                  fontSize: 10,
+                  fontWeight: FontWeight.bold),
+            ),
+          ],
+        ),
+      ],
+    ),
+  );
 }
